requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
# Django==4.2
django-comment-system==2.10.4
django-rating-system==1.1.1
django-reaction-system==1.1.1
idna==3.4
Pillow==9.5.0
pyTelegramBotAPI==4.10.0
python-decouple==3.8
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
telebot==0.0.5
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
